Isolate experiment components
